+++ src/main/java/ch/bitagent/bitcoin/lib/network/Electrum.java
@@ -1,10 +1,13 @@
 package ch.bitagent.bitcoin.lib.network;
 
+import ch.bitagent.bitcoin.lib.block.Block;
+import ch.bitagent.bitcoin.lib.ecc.Hex;
 import ch.bitagent.bitcoin.lib.helper.Helper;
 import ch.bitagent.bitcoin.lib.helper.Properties;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
+import java.io.ByteArrayInputStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Logger;
@@ -166,13 +169,26 @@ public class Electrum {
         return json.getJSONArray("result");
     }
 
-    public Integer height() {
+    public JSONObject headers() {
         var jsonResponse = callSocket(defaultSocket, getJsonRequest("blockchain.headers.subscribe", null));
         if (jsonResponse == null) {
             return null;
         }
         var json = new JSONObject(jsonResponse);
-        return json.getJSONObject("result").getInt("height");
+        var result = json.getJSONObject("result");
+
+        var height = result.getInt("height");
+
+        var hexString = result.getString("hex");
+        var hex = Hex.parse(hexString);
+        var stream = new ByteArrayInputStream(hex.toBytes());
+        var block = Block.parse(stream);
+        int timestamp = block.getTimestamp().intValue();
+
+        var jsonResult = new JSONObject();
+        jsonResult.put("height", height);
+        jsonResult.put("timestamp", timestamp);
+        return jsonResult;
     }
 
     public JSONArray getHistory(String scripthash) {